backend/requirements.txt
flask
flask-cors
openai
python-dotenv
langchain-openai
langgraph
tavily-python
requests